Get the legends for the item columns for CDI forms, with which category they belong to.

Get the legends for the item columns for CDI forms, with which category they belong to.

Usage

get_cdi_dict(form = c("WG", "WS"))

Arguments

form

Which kind of cdi form is this (WG or WS)?

Details

Items include the vocabulary checklist items, gesture items, and sentence items. These two csv (for WG and WS) was downloaded from WebCDI's GitHub repo, specifically the files [English_WG].csv and [English_WS].csv.