Load VIHI annotation data

Clone BLAB-private vihi_annotations repo to ~/BLAB_DATA once before using this function.

get_vihi_annotations(
  version = NULL,
  subset = c("random", "everything", "VI+TD-VI"),
  table = c("annotations", "intervals", "merged", "all"),
  include_all_tier_types = FALSE,
  allow_annotation_errors = FALSE,
  include_pi = FALSE
)

Arguments

version

version tag to checkout

subset

Which pre-defined subset of the data should be loaded?

'random' (the default) loads the annotations from the 15 randomly sampled intervals from all recordings in the corpus.

'VI+TD-VI' loads the annotations from the random and the top-5 high-volubility intervals from VI recordings and their TD matches.

'everything' loads all annotations from all tiers. Exercise caution with this option: the data will include incomplete and unchecked annotations.

table

Which table to return - annotations (the default) or intervals. If merged, returns the annotations table with the interval information merged in. Intervals without annotations won't be included. If all, returns a named list of both tables.#'

include_all_tier_types

Should all tier types be included in the output? If FALSE (the default), only tiers that are relevant to the subset are returned. For the 'random' and 'VI+TD-VI' subsets, the relevant tier types are: transcription, vcm, lex, mwu, xds. For the 'everything' subset, this parameter is ignored as all tier types are returned.

allow_annotation_errors

In case errors are found in the annotations, should the function throw an error (FALSE, the default) or add error_n columns to the annotations table? Use only as a way to inspect the errors, not as a way to ignore them.

include_pi

Should annotations marked as PI be included in the output? If FALSE (the default), they are filtered out.

Value

A table or a list of tables depending on the table parameter.

Details

The speaker TIER is identified by the participant column. Other tiers are in columns.

Notes:

Annotation are checked for errors for the standard ACLEW tiers only. Interval-level checks aren't currently checked at all.

Annotations marked as PI are included. Filter them out if you don't want them.

The transcribed utterance can be empty (”). Normally, that means that a code interval has been segmented but not annotated. But there might be other stray utterance segments like that.

(relevant for non-speaker TIERs only) Currently, there is no way to tell whether an annotation is missing because it was not segmented or because it was segmented but not yet annotated: both are represented as NA. This will change in the future: missing segment will still be NA, but missing annotation will be ”.

Examples

vitd_annotations <- get_vihi_annotations(version='0.0.0.9006-dev.5',
                                         subset='VI+TD-VI')
#> You've requested version 0.0.0.9006-dev.5 of vihi_annotations.
#> The latest available version is 0.0.0.9010.
#> Consider updating.
#> reading file: /Users/kiennguyen/BLAB_DATA/vihi_annotations/annotations.csv
#> reading file: /Users/kiennguyen/BLAB_DATA/vihi_annotations/intervals.csv
#> reading file: /Users/kiennguyen/BLAB_DATA/vihi_annotations/vi_td_matches.csv

vitd <- get_vihi_annotations(version='0.0.0.9006-dev.5', subset='VI+TD-VI',
                             table='all')
#> You've requested version 0.0.0.9006-dev.5 of vihi_annotations.
#> The latest available version is 0.0.0.9010.
#> Consider updating.
#> reading file: /Users/kiennguyen/BLAB_DATA/vihi_annotations/annotations.csv
#> reading file: /Users/kiennguyen/BLAB_DATA/vihi_annotations/intervals.csv
#> reading file: /Users/kiennguyen/BLAB_DATA/vihi_annotations/vi_td_matches.csv
vitd$annotations %>% head()
#> # A tibble: 6 × 12
#>   eaf_filename   participant   onset offset transcription transcription_id mwu  
#>   <chr>          <chr>         <int>  <int> <chr>         <chr>            <chr>
#> 1 TD_436_678.eaf MC2         4747362 4.75e6 xxx you know… a489             NA   
#> 2 TD_436_678.eaf MC1         4750138 4.75e6 xxx my ball!  a481             NA   
#> 3 TD_436_678.eaf MC1         4754066 4.76e6 Dad!          a482             NA   
#> 4 TD_436_678.eaf MA1         4756245 4.76e6 what?         a366             NA   
#> 5 TD_436_678.eaf MC1         4756548 4.76e6 he took my b… a483             NA   
#> 6 TD_436_678.eaf MC1         4760794 4.76e6 he took my b… a484             NA   
#> # ℹ 5 more variables: lex <chr>, vcm <chr>, xds <chr>, code_num <chr>, PI <chr>
vitd$intervals %>% head()
#> # A tibble: 6 × 10
#>   eaf_filename   code_num sampling_type   is_silent   onset offset context_onset
#>   <chr>          <chr>    <chr>           <chr>       <int>  <int>         <int>
#> 1 TD_436_678.eaf 1        random          N         4740000 4.86e6       4620000
#> 2 TD_436_678.eaf 19       high-volubility N         6077000 6.20e6       5957000
#> 3 TD_436_678.eaf 20       high-volubility N         6797000 6.92e6       6677000
#> 4 TD_436_678.eaf 2        random          N         7980000 8.10e6       7860000
#> 5 TD_436_678.eaf 3        random          N         8340000 8.46e6       8220000
#> 6 TD_436_678.eaf 4        random          N         9720000 9.84e6       9600000
#> # ℹ 3 more variables: context_offset <int>, rank <int>, is_top_5_high_vol <lgl>